@@ -11,33 +11,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.13",
-    "@angular/common": "^7.2.13",
-    "@angular/compiler": "^7.2.13",
-    "@angular/core": "^7.2.13",
-    "@angular/forms": "^7.2.13",
-    "@angular/http": "^7.2.13",
-    "@angular/platform-browser": "^7.2.13",
-    "@angular/platform-browser-dynamic": "^7.2.13",
-    "@angular/router": "^7.2.13",
-    "@ng-bootstrap/ng-bootstrap": "^4.1.1",
+    "@angular/animations": "^8.2.14",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "@types/chart.js": "^2.7.54",
     "chart.js": "^2.8.0",
     "core-js": "^2.5.4",
     "jwt-decode": "^2.2.0",
-    "rxjs": "^6.0.0",
+    "rxjs": "^6.6.3",
     "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "~7.3.8",
-    "@angular/compiler-cli": "^7.2.13",
-    "@angular/language-service": "^7.2.13",
+    "@angular-devkit/build-angular": "~0.803.29",
+    "@angular/cli": "~8.3.29",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~1.7.1",
@@ -48,6 +46,6 @@
     "protractor": "~5.3.0",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~3.2.4"
+    "typescript": "~3.5.3"
   }
 }
